solver/requirements.txt
sympy
docopt-ng
jinja2

